[#22] 🩹테스트 실패 수정

+++ domain/src/test/kotlin/com/splab/springgames/domain/member/vo/GameCardTotalPriceTest.kt
@@ -18,7 +18,7 @@ class GameCardTotalPriceTest : DescribeSpec({
                 val gameCardTotalPrice = GameCardTotalPrice(value.toBigDecimal())
 
                 // assert
-                gameCardTotalPrice.value shouldBe value
+                gameCardTotalPrice.value shouldBe value.toBigDecimal()
             }
         }
